Refactor search station contract, presenter for fixing bug
In station search presenter, refactor filter stations in province method to filter as station name and localized lowercase instead of province with standart lowercase method for fixing stations not showing up after filtering in station name bug and refactor did set stations in province method to show only 3 of the stations if there is more. Refactor search station contract and related presenter and interactor for fixing bug of stations in province not getting back when search text is deleted back to empty. Refactor view layer to add dismissing the keyboard when return button pressed in the keyboard and refactor table view helper to deselect the row of table view after selecting it